backend/requirements.txt
# =============================================================================
# Django Core
# =============================================================================
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.0.0

# =============================================================================
# Authentication
# =============================================================================
djangorestframework-simplejwt>=5.2.0

# =============================================================================
# Task Queue & Scheduling
# =============================================================================
celery>=5.3.0
django-celery-beat>=2.6.0
redis>=5.0.0
django-redis>=5.4.0

# =============================================================================
# Rate Limiting
# =============================================================================
django-ratelimit>=4.1.0

# =============================================================================
# Email
# =============================================================================
redmail>=0.6.0
Jinja2>=3.1.0

# =============================================================================
# Resume Parsing - PDF / DOCX Extraction
# =============================================================================
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.17
pymupdf-layout>=0.0.1
pdfplumber>=0.10.0
pypdf>=3.0.0
python-docx>=1.0.0

# =============================================================================
# Text Processing & NLP
# =============================================================================
# spaCy requires Visual Studio C++ Build Tools on Windows
# Install: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# spacy>=3.7.0
regex>=2023.0.0
textblob>=0.18.0

# =============================================================================
# Date & Contact Parsing
# =============================================================================
python-dateutil>=2.8.0
phonenumbers>=8.13.0
email-validator>=2.1.0

# =============================================================================
# ML / AI — Embeddings & Scikit-learn
# =============================================================================
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
joblib>=1.3.0

# =============================================================================
# LLM Clients
# =============================================================================
openai>=1.0.0
groq>=0.9.0

# =============================================================================
# LangChain
# =============================================================================
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-groq>=0.1.0

# =============================================================================
# Phase 2 — Certification Engine
# =============================================================================
weasyprint==62.3
qrcode[pil]==7.4.2
Pillow>=10.0.0

# =============================================================================
# Data & Reporting Utilities
# =============================================================================
openpyxl>=3.1.0
requests>=2.31.0

# =============================================================================
# General Utilities
# =============================================================================
pytz>=2023.3
